requirements.txt
flask
boto3
pymysql
sqlalchemy
python-dotenv
cryptography

